+++ R/calculations.R
@@ -66,10 +66,6 @@ calc_obs_props <- function(obj, dataset, cols = NULL, keep_other_cols = TRUE) {
 }
 
 
-
-
-
-
 #' Compare treatments
 #' 
 #' Apply a function to compare data, usually abundance, from pairs of 
@@ -221,3 +217,61 @@ compare_treatments <- function(obj, dataset, sample_ids, treatments,
   # Convert to tibble and return
   dplyr::as.tbl(output)
 }
+
+
+#' Sum observation values for each taxon
+#' 
+#' For a given table in a taxmap object, sum the values in each column for each
+#' taxon. This is useful to convert per-observation counts (e.g. OTU counts) to
+#' per-taxon counts.
+#' 
+#' @param obj A taxmap object
+#' @param dataset The name of a table in \code{obj} that contains counts.
+#' @param cols The names/indexes of columns in \code{data} that have counts. By
+#'   Default, all numeric columns in \code{data} are used.
+#'   
+#' @return A tibble
+#'   
+#' @export
+calc_taxon_abund <-function(obj, dataset, cols = NULL) {
+  # Check that dataset exists and is a table
+  if (! dataset %in% names(obj$data)) {
+    stop(paste0('The dataset "', dataset,
+                '" is not in the object supplied. Datasets found include:\n  ',
+                limited_print(names(obj$data), type = "silent")))
+  }
+  if (! is.data.frame(obj$data[[dataset]])) {
+    stop(paste0('The dataset "', dataset,  '" is not a table.'))
+  }
+  
+  # Get count table
+  count_table <- obj$data[[dataset]]
+  
+  # Find default columns if needed
+  if (is.null(cols)) {
+    cols <- which(vapply(count_table, is.numeric, logical(1)))
+  }
+  
+  # Check that count columns are numeric
+  col_is_num <- vapply(count_table[cols], is.numeric, logical(1))
+  if (! all(col_is_num)) {
+    stop(paste0("All columns must be numeric. The following columns are not numeric:  ",
+                limited_print(cols[!col_is_num], type = "silent")))
+  }
+  
+  # Sum counts per taxon for each sample
+  output <- lapply(cols, function(col_index) {
+    count_table[col_index]
+    obs_apply(obj, dataset, function(i) sum(count_table[i, col_index]), simplify = TRUE)
+  })
+  output <- as.data.frame(output, stringsAsFactors = FALSE)
+  colnames(output) <- colnames(count_table[cols])
+  
+  # Add taxon_id column
+  output <- cbind(data.frame(taxon_id = obj$taxon_ids(), stringsAsFactors = FALSE),
+                  output)
+  
+  # Convert to tibble and return
+  dplyr::as.tbl(output)
+}
+